Resolved comments of localStorage and helpers
