+++ src/root/AdminRoutes.jsx
@@ -1,39 +1,17 @@
-import {
-  Redirect,
-  Route,
-  Switch,
-  useHistory,
-  useRouteMatch,
-} from 'react-router-dom';
+import { Redirect, Route, Switch, useRouteMatch } from 'react-router-dom';
 
 import ProtectedAdminRoute from 'HOC/ProtectedAdminRoute';
 import LoginContainer from 'modules/admin/login/LoginContainer';
 import HomeContainer from 'modules/admin/home/HomeContainer';
 
 import { ADMIN_ROUTES } from 'constants/routeConstants';
-import { ADMIN_LOGIN } from 'redux/admin/login/actionConstants';
 
-import axios from 'axios';
-import { useDispatch } from 'react-redux';
+import { adminRoutesHelper } from 'modules/admin/AdminRoutesHelper';
 
 function AdminRoutes() {
   const { path } = useRouteMatch();
+  adminRoutesHelper();
 
-  const history = useHistory();
-  const dispatch = useDispatch();
-  axios.interceptors.response.use(
-    (res) => res,
-    (err) => {
-      if (err.response.status === 401) {
-        dispatch({
-          type: ADMIN_LOGIN.ADMIN_FAILURE_ACTION,
-          payload: 'Unauthorized User',
-        });
-        history.push('/admin/login');
-      }
-      throw err;
-    },
-  );
   return (
     <Switch>
       <Route
